refactor: move 'prisma' module to 'infra' directory
