+++ src/domain/user/user.module.ts
@@ -3,7 +3,7 @@ import { UserControllerV1 } from "./v1/user.controller";
 import { UserControllerV2 } from "./v2/user.controller";
 import { UserService } from "./user.service";
 import { UserRepository } from "./user.repository";
-import { PrismaModule } from "src/prisma/prisma.module";
+import { PrismaModule } from "src/infra/prisma/prisma.module";
 import { CommonModule } from "src/common/common.module";
 
 @Module({
