+++ test/helpers/test-database.helper.ts
@@ -1,4 +1,4 @@
-import { PrismaService } from "src/prisma/prisma.service";
+import { PrismaService } from "src/infra/prisma/prisma.service";
 
 interface IUserCredentials {
   name: string;
